webapp/requirements.txt
attrs==23.1.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.0.7
CouchDB==1.2
cycler==0.11.0
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Fiona==1.9.4
Flask==2.2.3
fonttools==4.39.3
geopandas==0.13.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
plotly==5.14.1
pyparsing==3.0.9
pyproj==3.5.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
requests==2.30.0
shapely==2.0.1
six==1.16.0
tenacity==8.2.2
text-unidecode==1.3
tomli==2.0.1
typing-extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.2.3
wordcloud==1.8.2.2
zipp==3.15.0
